refactor(user): FormLayout을 RootLayout을 사용하여 리팩토링
