+++ apps/user/src/app/faq/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { CategoryFilter, FaqList } from '@/components/faq';
-import { AppLayout } from '@/layouts';
+import { Layout } from '@/layouts';
 import { color } from '@maru/theme';
 import { Loader, Text } from '@maru/ui';
 import { Suspense, useState } from 'react';
@@ -11,7 +11,7 @@ const FaqPage = () => {
     const [category, setCategory] = useState('TOP_QUESTION');
 
     return (
-        <AppLayout header footer>
+        <Layout header footer>
             <StyledFaqPage>
                 <Text fontType="H1" color={color.gray900}>
                     자주 묻는 질문
@@ -21,7 +21,7 @@ const FaqPage = () => {
                     <FaqList category={category} />
                 </Suspense>
             </StyledFaqPage>
-        </AppLayout>
+        </Layout>
     );
 };
