+++ apps/user/src/app/form/전형선택/전형선택.tsx
@@ -1,10 +1,10 @@
 import { FormController } from '@/components/form';
-import { FormLayout } from '@/layouts';
 import { useFormStore } from '@/store';
 import { Column, Radio, Row, Td, Th } from '@maru/ui';
 import { flex } from '@maru/utils';
 import { useEffect } from 'react';
 import { styled } from 'styled-components';
+import FormLayout from '../layout';
 import { useCTAButton, useInput } from './전형선택.hooks';
 
 const 전형선택 = () => {
@@ -15,7 +15,9 @@ const 전형선택 = () => {
     useEffect(() => {
         if (form.education.graduationType === 'QUALIFICATION_EXAMINATION') {
             if (form.type !== 'REGULAR') {
-                alert('서류상으로 검정고시 합격자는 특별전형 지원이 불가능해요. 일반전형으로 지원해주세요!');
+                alert(
+                    '서류상으로 검정고시 합격자는 특별전형 지원이 불가능해요. 일반전형으로 지원해주세요!',
+                );
                 setForm((prev) => ({ ...prev, type: 'REGULAR' }));
             }
         }
