+++ apps/user/src/layouts/index.ts
@@ -1,2 +1,1 @@
-export { default as AppLayout } from './AppLayout';
-export { default as FormLayout } from './FormLayout';
+export { default as Layout } from './Layout';
